linting ErrorBoundary
